refactor(web): split OnboardingPage into onboarding step modules (#3712)
Extract pure helpers, types, static content, storage utilities and the seven onboarding step render blocks (choose, comfort, privacy, newcomer, goals, template, complete) into dedicated modules under src/pages/onboarding/. OnboardingPage.tsx drops from ~2,252 to ~960 lines while behavior is unchanged (35 page tests still pass).

Also finalize normalizeNegativeFormat to return NegativeFormat so display-settings typechecks cleanly (#3283).
